Adjust segfault checker for more precise resets

+++ check_for_tskloaddb_segfault.sh
@@ -5,24 +5,37 @@
 #set -e #Don't disable error-halts.
 set -x
 
-test -r DRIVE2_TIME_FINAL.E01
+IMAGE=../../DRIVE2_TIME_FINAL.E01
+if [ ! -r "$IMAGE" ]; then
+  echo "ERROR:$(basename $0):Soft-link the disk image into the test superproject." >&2
+  exit 1
+fi
 
 #Cleanup from last test
 rm -rf build test.db
-rm -rf tsk3 #An artifact of checkouts
-git checkout -- .
+if [ -d tsk3 ]; then
+  rm -rf tsk3 #An artifact of checkouts
+  git checkout -- tsk3
+fi
 
 ./bootstrap
 ./configure --disable-java --prefix=$PWD/build
 make -j || exit 125   #We're looking for a segfault, not a build failure.
 make install
 
-#Segfault here -> 'bad' commit.  Any other execution = 'good' commit, for the purpose of this test.
-build/bin/tsk_loaddb -d test.db DRIVE2_TIME_FINAL.E01
+build/bin/tsk_loaddb -d test.db "$IMAGE"
 rc=$?
 
 echo "NOTE:$(basename $0):Resetting tsk3/ dir (if it's there) for next git-bisect checkout." >&2
-rm -rf tsk3
-git checkout -- .
 
-exit $rc
+if [ -d tsk3 ]; then
+  rm -rf tsk3
+  git checkout -- tsk3
+fi
+
+#Segfault from tsk_loaddb -> 'bad' commit.  Any other exit status -> 'good' commit, for the purpose of this test.
+if [ $rc -eq 139 ]; then
+  exit 1
+else
+  exit 0
+fi
